+++ crates/bevy_mesh/Cargo.toml
@@ -17,7 +17,7 @@ bevy_math = { path = "../bevy_math", version = "0.18.0-dev" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.18.0-dev" }
 bevy_ecs = { path = "../bevy_ecs", version = "0.18.0-dev" }
 bevy_transform = { path = "../bevy_transform", version = "0.18.0-dev" }
-bevy_mikktspace = { version = "0.17.0-dev" }
+bevy_mikktspace = { version = "0.17.0-dev", optional = true }
 bevy_derive = { path = "../bevy_derive", version = "0.18.0-dev" }
 bevy_platform = { path = "../bevy_platform", version = "0.18.0-dev", default-features = false, features = [
   "std",
