+++ crates/bevy_mesh/src/lib.rs
@@ -7,6 +7,7 @@ mod components;
 mod conversions;
 mod index;
 mod mesh;
+#[cfg(feature = "bevy_mikktspace")]
 mod mikktspace;
 #[cfg(feature = "morph")]
 pub mod morph;
@@ -20,6 +21,7 @@ use bitflags::bitflags;
 pub use components::*;
 pub use index::*;
 pub use mesh::*;
+#[cfg(feature = "bevy_mikktspace")]
 pub use mikktspace::*;
 pub use primitives::*;
 pub use vertex::*;
